requirements.txt
black==25.1.0
click==8.2.1
mypy==1.17.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
typing_extensions==4.14.1
